class8/requirements.txt
# AI Safety Demonstrations - Required Dependencies
# Core libraries for AI safety analysis
openai>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0

# Testing and development
pytest>=7.0.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Text processing and analysis
nltk>=3.8.0
textblob>=0.17.0
scikit-learn>=1.3.0

# Data validation and safety
pydantic>=2.0.0
validators>=0.20.0

# Security and safety utilities  
bleach>=6.0.0
cryptography>=41.0.0

# Optional: For advanced bias detection
fairlearn>=0.9.0
aif360>=0.5.0

# Optional: For web interface (if creating interactive demo)
streamlit>=1.25.0
gradio>=3.40.0 